test(hooks): split runtime fast-fail tests into their own module
The hardened membership pins pushed test_lefthook_prepush_fast_fail.py
to 570 lines, over the 500-line file-size taste rule that the fast stage
itself enforces. Split along the structural/runtime seam: config pins
stay put, the fixture-repo runtime tests move to
test_lefthook_prepush_fast_fail_runtime.py unchanged.

Refs #5066
